==> code_data/1_Data_Acquisition_CAISO_Prices_Load_Fcasts.ipynb <==
---

## Import modules and classes

import wget, os, zipfile
import time

---

## Functions to Construct Queries

(for Accessing CAISO OASIS API)

# Function to construct a one-month electricity Day-Ahead-Market
#    data query for the CAISO OASIS API

# Parameters (all strings):
#    node                       name of node
#    startyear, startmonth


def create_DAM_query(node, startyear, startmonth):
    oasis_website = 'oasis.caiso.com'
    context_path  = 'oasisapi'

    url = f'http://{oasis_website}/{context_path}/SingleZip'

    resultformat  =  '6'
    queryname     =  'PRC_LMP'
    version       =  '1'
    market_run_id =  'DAM'
    
    if startmonth == '12':
        endmonth  =  '01'
        endyear   =  f'{int(startyear) + 1}'
    else:
        endmonth  =  f'{(int(startmonth) + 1):02d}'
        endyear   =  startyear

    startdatetime =  f'{startyear}{startmonth}01T08:00-0000'
    enddatetime   =  f'{endyear}{endmonth}01T07:00-0000'

    query = f'{url}?resultformat={resultformat}&queryname={queryname}\
&version={version}&startdatetime={startdatetime}&enddatetime={enddatetime}\
&market_run_id={market_run_id}&node={node}'
    
    return query

# Function to construct a one-month electricity Hour-Ahead-Scheduling-Process
#    data query for the CAISO OASIS API

# Parameters (all strings):
#    node                       name of node
#    startyear, startmonth


def create_HASP_query(node, startyear, startmonth):
    oasis_website = 'oasis.caiso.com'
    context_path  = 'oasisapi'

    url = f'http://{oasis_website}/{context_path}/SingleZip'

    resultformat  =  '6'
    queryname     =  'PRC_HASP_LMP'
    version       =  '3'
    market_run_id =  'HASP'
    
    if startmonth == '12':
        endmonth  =  '01'
        endyear   =  f'{int(startyear) + 1}'
    else:
        endmonth  =  f'{(int(startmonth) + 1):02d}'
        endyear   =  startyear

    startdatetime =  f'{startyear}{startmonth}01T08:00-0000'
    enddatetime   =  f'{endyear}{endmonth}01T07:00-0000'

    query = f'{url}?resultformat={resultformat}&queryname={queryname}\
&version={version}&startdatetime={startdatetime}&enddatetime={enddatetime}\
&market_run_id={market_run_id}&node={node}'
    
    return query

# Function to construct a GENERIC one-month .csv LMP price download
#    data query for the CAISO OASIS API

# Parameters (all strings):
#    node                       name of node
#    startyear, startmonth


def price_query(queryname, version, startyear, startmonth, market_run_id, node):
    oasis_website = 'oasis.caiso.com'
    context_path  = 'oasisapi'

    url = f'http://{oasis_website}/{context_path}/SingleZip'

    resultformat  =  '6'
    
    if startmonth == '12':
        endmonth  =  '01'
        endyear   =  f'{int(startyear) + 1}'
    else:
        endmonth  =  f'{(int(startmonth) + 1):02d}'
        endyear   =  startyear

    startdatetime =  f'{startyear}{startmonth}01T08:00-0000'
    enddatetime   =  f'{endyear}{endmonth}01T07:00-0000'

    query = f'{url}?resultformat={resultformat}&queryname={queryname}&version={version}\
&startdatetime={startdatetime}&enddatetime={enddatetime}\
&market_run_id={market_run_id}&node={node}'
    
    return query

# Function to construct a GENERIC one-month .csv load (demand) forecast
#    download data query for the CAISO OASIS API

# Parameters (all strings):
#    startyear, startmonth


def load_query(queryname, version, startyear, startmonth, market_run_id):
    oasis_website = 'oasis.caiso.com'
    context_path  = 'oasisapi'

    url = f'http://{oasis_website}/{context_path}/SingleZip'

    resultformat  =  '6'
    
    if startmonth == '12':
        endmonth  =  '01'
        endyear   =  f'{int(startyear) + 1}'
    else:
        endmonth  =  f'{(int(startmonth) + 1):02d}'
        endyear   =  startyear

    startdatetime =  f'{startyear}{startmonth}01T08:00-0000'
    enddatetime   =  f'{endyear}{endmonth}01T07:00-0000'

    query = f'{url}?resultformat={resultformat}&queryname={queryname}&version={version}\
&market_run_id={market_run_id}&startdatetime={startdatetime}&enddatetime={enddatetime}'
    
    return query

---

## Get Energy Price Zipfiles 

(from CAISO OASIS website via API)

node = '0096WD_7_N001'

for i in range(1, 12):
    if i % 12 == 0:
        startyear  = str(2019 + i//12 - 1)
        startmonth = f'{12:02d}'
    else:
        startyear  = str(2016 + i//12)
        startmonth = f'{i%12:02d}'
    
    wget.download(price_query('PRC_LMP',
                              '3',
                              startyear,
                              startmonth,
                              'DAM',
                              node), '../raw_data/caiso_downloads/caiso_dam_dl/')
    time.sleep(5)
    
    wget.download(price_query('PRC_HASP_LMP',
                              '3',
                              startyear,
                              startmonth,
                              'HASP',
                              node), '../raw_data/caiso_downloads/caiso_hasp_dl/')
    time.sleep(5)
    
    wget.download(price_query('PRC_INTVL_LMP',
                              '3',
                              startyear,
                              startmonth,
                              'RTM',
                              node), '../raw_data/caiso_downloads/caiso_rtm_dl/')
    time.sleep(5)

for i in range(1, 12):
    if i % 12 == 0:
        startyear  = str(2019 + i//12 - 1)
        startmonth = f'{12:02d}'
    else:
        startyear  = str(2019 + i//12)
        startmonth = f'{i%12:02d}'
    
    wget.download(load_query('SLD_FCST',
                             '1',
                             startyear,
                             startmonth,
                             '7DA',
                            ), '../raw_data/caiso_downloads/caiso_7da_load_dl/')
    time.sleep(5)    
    
    wget.download(load_query('SLD_FCST',
                             '1',
                             startyear,
                             startmonth,
                             '2DA',
                            ), '../raw_data/caiso_downloads/caiso_2da_load_dl/')
    time.sleep(5)    
    
    
    wget.download(load_query('SLD_FCST',
                             '1',
                             startyear,
                             startmonth,
                             'DAM',
                            ), '../raw_data/caiso_downloads/caiso_dam_load_dl/')
    time.sleep(5)    
    
    wget.download(load_query('SLD_FCST',
                             '1',
                             startyear,
                             startmonth,
                             'RTM',
                            ), '../raw_data/caiso_downloads/caiso_rtm_load_dl/')
    time.sleep(5)    

# Single month download cell

caiso_dam_dl =  '../raw_data/caiso_downloads/caiso_dam_dl/'
caiso_hasp_dl = '../raw_data/caiso_downloads/caiso_hasp_dl/'

node = '0096WD_7_N001'

startyear  = '2019'
startmonth = '05'

wget.download(create_DAM_query(node, startyear, startmonth), caiso_dam_dl)
time.sleep(10)
wget.download(create_HASP_query(node, startyear, startmonth), caiso_hasp_dl)

---

## Unzip Downloaded CAISO Files

(to .csv Directories)

unzipped_caiso_dam = '../raw_data/unzipped_caiso/unzipped_caiso_dam/'
unzipped_caiso_hasp = '../raw_data/unzipped_caiso/unzipped_caiso_hasp/'


caiso_dam_dl = '../raw_data/caiso_downloads/caiso_dam_dl/'
caiso_hasp_dl = '../raw_data/caiso_downloads/caiso_hasp_dl/'


for item in os.listdir(caiso_dam_dl):           # loop through items in dir
    if item.split('.')[-1] == 'zip':            # check for zip extension
        file_name = f'{caiso_dam_dl}{item}'     # get relative path of files
        print(f'unzipping... {file_name}')
        zip_ref = zipfile.ZipFile(file_name)    # create zipfile object
        with zip_ref as target:
            target.extractall(unzipped_caiso_dam)
    else: continue

for item in os.listdir(caiso_hasp_dl):           # loop through items in dir
    if item.split('.')[-1] == 'zip':            # check for zip extension
        file_name = f'{caiso_hasp_dl}{item}'     # get relative path of files
        print(f'unzipping... {file_name}')
        zip_ref = zipfile.ZipFile(file_name)    # create zipfile object
        with zip_ref as target:
            target.extractall(unzipped_caiso_hasp)
    else: continue

download_dir = '../raw_data/caiso_downloads/'
unzipped_target_dir = '../raw_data/unzipped_caiso/unzipped_caiso_rtm/'

for item in os.listdir(download_dir):           # loop through items in dir
    if item.split('.')[-1] == 'zip':            # check for zip extension
        file_name = f'{download_dir}{item}'     # get relative path of files
        print(f'unzipping... {file_name}')
        zip_ref = zipfile.ZipFile(file_name)    # create zipfile object
        with zip_ref as target:
            target.extractall(unzipped_target_dir)
    else: continue

def unzip_dir(download_dir, unzipped_target_dir):
    for item in os.listdir(download_dir):           # loop through items in dir
        if item.split('.')[-1] == 'zip':            # check for zip extension
            file_name = f'{download_dir}{item}'     # get relative path of files
            print(f'unzipping... {file_name}')
            zip_ref = zipfile.ZipFile(file_name)    # create zipfile object
            with zip_ref as target:
                target.extractall(unzipped_target_dir)
        else: continue

unzip_dir('../raw_data/caiso_downloads/caiso_7da_load_dl/',
          '../raw_data/unzipped_caiso/unzipped_caiso_7da_load/')

unzip_dir('../raw_data/caiso_downloads/caiso_2da_load_dl/',
          '../raw_data/unzipped_caiso/unzipped_caiso_2da_load/')

unzip_dir('../raw_data/caiso_downloads/caiso_dam_load_dl/',
          '../raw_data/unzipped_caiso/unzipped_caiso_dam_load/')

unzip_dir('../raw_data/caiso_downloads/caiso_rtm_load_dl/',
          '../raw_data/unzipped_caiso/unzipped_caiso_rtm_load/')
