==> code_data/2_Data_Acquisition_CA_Water_Levels.ipynb <==
---

## Import modules and classes

import wget, os, zipfile
import time

---

## Functions to Construct Queries

(for Accessing CAISO OASIS API)

# Function to construct a GENERIC one-month .csv download query
#    for hourly reservoir storage amounts from the CA DWR API

# Parameter (string):  reservoir alphabetic code


def res_stg_query(station):
    url = f'http://cdec.water.ca.gov/dynamicapp/req/CSVDataServlet'
    Stations = station
    SensorNums = '15'
    dur_code = 'H'
    Start =  '2016-01-01'
    End   =  '2019-05-07'

    query = f'{url}?Stations={station}&SensorNums={SensorNums}\
&dur_code={dur_code}&Start={Start}&End={End}'
    
    return query

res_stg_query('CLE')

---

## Get Hourly Reservoir Storage Sensor Readings

(in acre-feet from CA DWR website via API)

reservoirs_list = ['CLE', 'WHI', 'LEW', 'WRS',
                   'COY', 'SHA', 'KES', 'ORO',
                   'ANT', 'FRD', 'DAV', 'BUL',
                   'ENG', 'FOL', 'UNV', 'LON',
                   'ICH', 'NAT', 'INV', 'BER',
                   'BLB', 'NHG', 'CMN', 'PAR',
                   'DON', 'BRD', 'TUL', 'NML',
                   'DNP', 'HTH', 'CHV', 'EXC',
                   'BUC', 'HID', 'MIL', 'SNL',
                   'PNF', 'TRM', 'SCC', 'ISB',
                   'STP', 'INP', 'DNN', 'CCH',
                   'PYM', 'CAS', 'PRR'
                  ]

len(reservoirs_list)

for station in reservoirs_list:
    wget.download(res_stg_query(station), f'../ca_dwr_dl/{station}_hrly_strg.csv')
    time.sleep(5)

# Single reservoir download

# station = 'CLE'
# wget.download(res_stg_query(station), f'../ca_dwr_dl/{station}_hrly_strg.csv')
