==> code_data/6_NOAA_Four_CA_Weather_Stn_Dataframes.ipynb <==
---

## Imports

import pandas as pd
import numpy as np

import wget, os
import time
import glob

import matplotlib.pyplot as plt
import seaborn as sns
from pandas.plotting import register_matplotlib_converters
register_matplotlib_converters()
%matplotlib inline
sns.set_style('whitegrid')

---

## Declare Variables, Directories, and Column Names

stations_dict = {'72290023188': 'sand',        # San Diego
                 '72286903171': 'rive',        # Riverside
                 '72592024257': 'redd',        # Redding
                 '72389093193': 'fres'         # Fresno
                }

stn_id_list   = list(stations_dict.keys())
stn_name_list = list(stations_dict.values())

print(f'NOAA Weather Station IDs:   {stn_id_list}')
print(f'NOAA Weather Station Names: {stn_name_list}')

# Dataframe names
sand_df = None
rive_df = None
redd_df = None
fres_df = None

download_dirs = ['../raw_data/noaa_weather/2016/',
                 '../raw_data/noaa_weather/2017/',
                 '../raw_data/noaa_weather/2018/',
                 '../raw_data/noaa_weather/2019/'
                ]

weather_orig_cols=['STATION',
                   'DATE',
                   'TMP',
                   'WND',
                   'CIG',
                   'VIS']

weather_new_cols =['stn_id',
                   'datetime',
                   'temp',
                   'wind',
                   'ceil',
                   'vis']

weather_rename_dict = {old: new for old, new in zip(weather_orig_cols,
                                                    weather_new_cols)}

---

##  Define Functions  

### Function to create NOAA weather station DataFrame

Parameters:  dataframe (df), index slice position in `stations_list`  

Function steps:  
 1. Instantiate DataFrame
 2. Loop through the 4 annual download directories
 3. Reading in data from .csv files, appending rows to new dataframe
 3. Sort by `datetime` column, cast to 'datetime' object, and set as index 
 4. Sort by new `datetime` index to show rows in chronological order
 
 
### Function to extract relevant statistics from NOAA column

Parameters:  dataframe (df), orig_col_name (string), new_col_name (string), slice_index  

Function steps:  
 1. Step through rows of a dataframe
 2. Extract a single statistic from a "cell" that is a tuple of strings
 3. Store that stat in a new column whose name is a parameter
 4. Drop the original data column from the dataframe
 
 
### Function to print relevant descriptive info on dataframe

Parameter:  dataframe (df)

Function steps:
 1. print df.shape
 2. print null counts
 3. print head(5)
 
 
 ### Function to replace  hard-coded "99xxx9" values w/ NaNs

Parameter:  dataframe (df)

Function steps:
 1. replace specific codes (depends on which column) NOAA uses as NaN
 2. return null counts

def create_stn_df(df, stn_idx):
    df       = pd.DataFrame(columns=weather_new_cols)
    stn_name = list(stations_dict.values())[stn_idx]
    stn_id   = list(stations_dict.keys())[stn_idx]

    for directory in download_dirs:
        file = directory + stn_id + '.csv'
        temp_df = pd.read_csv(file,
                              usecols=weather_orig_cols).rename(index=str,
                                                                columns=weather_rename_dict)
        df = df.append(temp_df, ignore_index=True, sort=True)

    df = df.sort_values(by='datetime').reset_index(drop=True)

    df['datetime'] = pd.to_datetime(df['datetime'])

    df.set_index('datetime', inplace=True)
    df.sort_index(inplace=True)
    print(f'{stn_name}_df DataFrame Created')
    
    return df

def extract_statistic(df, orig_col_name, new_col_name, slice_index):
    for row in range(0, df.shape[0]):
        try:
            statistic = df.loc[df.index[row], orig_col_name].split(',')[slice_index]
            df.loc[df.index[row], new_col_name] = statistic
            
        except IndexError: break
        except:
            df.drop(labels=df.index[row], axis=0, inplace=True)  

    df.drop(columns=[orig_col_name], inplace=True)

    return

def summarize(df):
    print(f'\nShape: {df.shape}')
    print(f'\n  Nulls:\n{df.isna().sum()}')
    print(f'\ndf.head()\n{df.head()}')
    
    return

def replace_nines(df):
    df.replace('9999', np.nan, inplace=True)
    df.replace('99999', np.nan, inplace=True)
    df.replace('999999', np.nan, inplace=True)
    df.isna().sum()
    
    return

---

## Read in Data and Create 4 Weather Station DataFrames

stn_name_list

sand_df = create_stn_df(sand_df, 0)
rive_df = create_stn_df(rive_df, 1)
redd_df = create_stn_df(redd_df, 2)
fres_df = create_stn_df(fres_df, 3)

---

## San Diego DataFrame

summarize(sand_df)

sand_df.drop_duplicates().shape

sand_df = sand_df.drop_duplicates()
sand_df.shape

extract_statistic(sand_df, 'temp', 'sand_temp', 0)
extract_statistic(sand_df, 'wind', 'sand_wind', 3)
extract_statistic(sand_df, 'vis',  'sand_vis',  0)
extract_statistic(sand_df, 'ceil', 'sand_ceil', 0)

sand_df.drop(columns=['stn_id'], inplace=True)

summarize(sand_df)

sand_df.dropna(inplace=True)
replace_nines(sand_df)
sand_df.isna().sum()

# decided not to fill, because they are at "odd" times
# that don't align with other measurent times (:51 past each hour)

# sand_df.sand_ceil.fillna(method='ffill', inplace=True)
# sand_df.isna().sum()

sand_df.dropna(inplace=True)

summarize(sand_df)

sand_df.to_csv('../data/san_diego_weather2.csv')

---

## Riverside DataFrame

summarize(rive_df)

rive_df.drop_duplicates().shape

rive_df = rive_df.drop_duplicates()
rive_df.shape

extract_statistic(rive_df, 'temp', 'rive_temp', 0)
extract_statistic(rive_df, 'wind', 'rive_wind', 3)
extract_statistic(rive_df, 'vis',  'rive_vis',  0)
extract_statistic(rive_df, 'ceil', 'rive_ceil', 0)

rive_df.drop(columns=['stn_id'], inplace=True)

summarize(rive_df)

rive_df.dropna(inplace=True)
replace_nines(rive_df)
rive_df.isna().sum()

# did not fill forward... small num of nulls

# rive_df.sand_COLUMN.fillna(method='ffill', inplace=True)
# rive_df.isna().sum()

rive_df.dropna(inplace=True)

summarize(rive_df)

rive_df.to_csv('../data/riverside_weather2.csv')

## Redding DataFrame

summarize(redd_df)

redd_df.drop_duplicates().shape

redd_df = redd_df.drop_duplicates()
redd_df.shape

extract_statistic(redd_df, 'temp', 'redd_temp', 0)
extract_statistic(redd_df, 'wind', 'redd_wind', 3)
extract_statistic(redd_df, 'vis',  'redd_vis',  0)
extract_statistic(redd_df, 'ceil', 'redd_ceil', 0)

redd_df.drop(columns=['stn_id'], inplace=True)

summarize(redd_df)

redd_df.dropna(inplace=True)
replace_nines(redd_df)
redd_df.isna().sum()

# decided not to fill, because they are at "odd" times
# that don't align with other measurent times (:53 past each hour)

# redd_df.sand_ceil.fillna(method='ffill', inplace=True)
# redd_df.isna().sum()

redd_df.dropna(inplace=True)

summarize(redd_df)

redd_df.to_csv('../data/redding_weather2.csv')

## Fresno DataFrame

summarize(fres_df)

fres_df.drop_duplicates().shape

fres_df = fres_df.drop_duplicates()
fres_df.shape

extract_statistic(fres_df, 'temp', 'fres_temp', 0)
extract_statistic(fres_df, 'wind', 'fres_wind', 3)
extract_statistic(fres_df, 'vis',  'fres_vis',  0)
extract_statistic(fres_df, 'ceil', 'fres_ceil', 0)

fres_df.drop(columns=['stn_id'], inplace=True)

summarize(fres_df)

fres_df.dropna(inplace=True)
replace_nines(fres_df)
fres_df.isna().sum()

# decided not to fill, because they are at "odd" times
# that don't align with other measurent times (:53 past each hour)

# fres_df.sand_COLUMN.fillna(method='ffill', inplace=True)
# fres_df.isna().sum()

fres_df.dropna(inplace=True)

summarize(fres_df)

fres_df.to_csv('../data/fresno_weather2.csv')
