==> code_data/8_NOAA_Combined_CA_Weather_Dataframe-(4_stn).ipynb <==
---

## Imports

import pandas as pd
import numpy as np

import wget, os
import time
import glob

---

## Function Definitions

def index_to_datetime(df):
    df['datetime'] = pd.to_datetime(df['datetime']).dt.round('H')
    df.set_index('datetime', inplace=True)
    df.sort_index(inplace=True)
    return df

def shapes_nulls():
    print(f'sand: {sand_df.shape[0]} rows, {sand_df.isna().sum().sum()} nulls')
    print(f'rive: {rive_df.shape[0]} rows, {rive_df.isna().sum().sum()} nulls')
    print(f'redd: {redd_df.shape[0]} rows, {redd_df.isna().sum().sum()} nulls')
    print(f'fres: {fres_df.shape[0]} rows, {fres_df.isna().sum().sum()} nulls')
    return

---

## Read in Four Individual Weather Station DataFrames

sand_df = pd.read_csv('../data/intermediate_stages/san_diego_weather2.csv')
rive_df = pd.read_csv('../data/intermediate_stages/riverside_weather2.csv')
redd_df = pd.read_csv('../data/intermediate_stages/redding_weather2.csv')
fres_df = pd.read_csv('../data/intermediate_stages/fresno_weather2.csv')

shapes_nulls()

sand_df = index_to_datetime(sand_df)
rive_df = index_to_datetime(rive_df)
redd_df = index_to_datetime(redd_df)
fres_df = index_to_datetime(fres_df)

sand_df.head()

rive_df.head()

redd_df.head()

fres_df.head()

---

## Join DataFrames

weather_df = sand_df
weather_df.shape

weather_df = weather_df.merge(fres_df,
                              how='outer',
                              left_index = True,
                              right_index = True)
weather_df.shape

weather_df = weather_df.merge(rive_df,
                              how='outer',
                              left_index = True,
                              right_index = True)
weather_df.shape

weather_df = weather_df.merge(redd_df,
                              how='outer',
                              left_index = True,
                              right_index = True)
weather_df.shape

weather_df.drop_duplicates(inplace=True)
weather_df.shape

weather_df = weather_df[~weather_df.index.duplicated(keep='first')]
weather_df.shape

datetime_index = pd.date_range(start = '2016-01-01 01:00',
                               end   = '2019-04-24 07:00',
                               freq  = 'H')
len(datetime_index)

weather_df = weather_df.reindex(datetime_index)
weather_df.fillna(method='ffill', inplace=True)

weather_df = weather_df.tz_localize('America/Los_Angeles',
                                    ambiguous=True,
                                    nonexistent='shift_forward')

weather_df.isna().sum().sum()

weather_df = weather_df.astype(int)

weather_df.shape

weather_df.head()

weather_df.to_csv('../data/ca_weather.csv')
