==> code_data/9_Fully_Consolidated_Dataframe.ipynb <==
---

## Imports

import pandas as pd
import numpy as np

import wget, os
import time
import glob
import pytz
import pickle

---

## Function Definitions

def index_to_datetime(df):
    df['datetime'] = pd.to_datetime(df['datetime'], utc=True).dt.round('H')
    df.set_index('datetime', inplace=True)
    df.sort_index(inplace=True)
    return df

def shapes_nulls():
    print(f'Elec Prices:     {elec_prices_df.shape[0]} rows, {elec_prices_df.isna().sum().sum()} nulls')
    print(f'Elec Dem Fcasts: {load_fcasts_df.shape[0]} rows, {load_fcasts_df.isna().sum().sum()} nulls')
    print(f'CA Water Levels: {water_levels_df.shape[0]} rows, {water_levels_df.isna().sum().sum()} nulls')
    print(f'CA Weather Data: {weather_df.shape[0]} rows, {weather_df.isna().sum().sum()} nulls')
    return

---

## Read in Four Sub- DataFrames

elec_prices_df  = pd.read_csv('../data/elec_prices_hrly.csv')
load_fcasts_df  = pd.read_csv('../data/elec_demand_hrly.csv')
water_levels_df = pd.read_csv('../data/ca_water_levels.csv')
weather_df      = pd.read_csv('../data/ca_weather.csv')

shapes_nulls()

---

## Set Datetime Indices

# pytz.all_timezones

elec_prices_df.rename(columns={'start_datetime': 'datetime'}, inplace=True)
elec_prices_df = index_to_datetime(elec_prices_df)
elec_prices_df = elec_prices_df.tz_convert('America/Los_Angeles')

load_fcasts_df.rename(columns={'start_datetime': 'datetime'}, inplace=True)
load_fcasts_df = index_to_datetime(load_fcasts_df)
load_fcasts_df = load_fcasts_df.tz_convert('America/Los_Angeles')

water_levels_df = index_to_datetime(water_levels_df)
water_levels_df = water_levels_df.tz_convert('America/Los_Angeles')

weather_df.rename(columns={'Unnamed: 0': 'datetime'}, inplace=True)
weather_df = index_to_datetime(weather_df)
weather_df = weather_df.tz_convert('America/Los_Angeles')

elec_prices_df.head()

load_fcasts_df.head()

water_levels_df.head()

weather_df.head()

---

## Join DataFrames

df = elec_prices_df
df.shape

df = df.merge(load_fcasts_df[['7da_load_fcast_mw',
                              '2da_load_fcast_mw',
                              'dam_load_fcast_mw',
                              'rtm_load_fcast_mw']],
              how='left',
              left_index = True,
              right_index = True)
df.shape

df = df.merge(water_levels_df,
              how='left',
              left_index = True,
              right_index = True)
df.shape

df = df.merge(weather_df,
              how='left',
              left_index = True,
              right_index = True)
df.shape

df.drop_duplicates(inplace=True)
df.shape

df = df[~df.index.duplicated(keep='first')]
df.shape

df.head()

df.fillna(method='ffill', inplace=True)
df.drop(index=df.index[0], inplace=True)
df.isna().sum().sum()

df.dtypes

df.shape

df.head()

df.to_csv('../data/consolidated_data.csv')

with open('../data/consolidated_data.pkl', 'wb') as f:
    pickle.dump(df, f)
